pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "editor_assistant"
version = "0.1"
dependencies = [
    "markitdown",
    "requests",
    "pydantic",
    "trafilatura",
    "readabilipy",
    "html2text",
    "pyyaml",
    "jinja2",
]

[project.scripts]
# Main CLI entry point
editor-assistant = "editor_assistant.src.editor_assistant.cli:main"

# Legacy entry points for backward compatibility  
generate-news = "editor_assistant.src.editor_assistant.cli:generate_news"
generate-outline = "editor_assistant.src.editor_assistant.cli:generate_outline"

# Utility commands (kept for compatibility)
any2md = "editor_assistant.src.editor_assistant.md_converter:main"
html2md = "editor_assistant.src.editor_assistant.clean_html_to_md:main"

[project.optional-dependencies]
dev = [
    "pytest",
]